+++ contracts/modules/treasury/ConfigManager.sol
@@ -11,32 +11,22 @@ import "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
 contract ConfigManager is Initializable, OwnableUpgradeable {
     // Configuration struct
     struct GameConfig {
-        uint ticketPrice;
-        uint gameDuration;
+        uint256 ticketPrice;
+        uint256 gameDuration;
         uint8 fee;
-        uint maxTicketsPerPlayer;
+        uint256 maxTicketsPerPlayer;
         bool isActive;
     }
 
     // Configuration mappings
     mapping(uint8 => GameConfig) public gameConfigs;
-    mapping(string => uint) public systemParams;
+    mapping(string => uint256) public systemParams;
     mapping(string => address) public contractAddresses;
 
     // Events
-    event GameConfigUpdated(
-        uint8 gameType,
-        uint ticketPrice,
-        uint gameDuration,
-        uint8 fee,
-        uint maxTickets
-    );
-    event SystemParamUpdated(string param, uint oldValue, uint newValue);
-    event ContractAddressUpdated(
-        string contractName,
-        address oldAddress,
-        address newAddress
-    );
+    event GameConfigUpdated(uint8 gameType, uint256 ticketPrice, uint256 gameDuration, uint8 fee, uint256 maxTickets);
+    event SystemParamUpdated(string param, uint256 oldValue, uint256 newValue);
+    event ContractAddressUpdated(string contractName, address oldAddress, address newAddress);
 
     /// @custom:oz-upgrades-unsafe-allow constructor
     constructor() {
@@ -62,13 +52,8 @@ contract ConfigManager is Initializable, OwnableUpgradeable {
         });
 
         // 7-days game config (gameType = 5)
-        gameConfigs[5] = GameConfig({
-            ticketPrice: 1 ether,
-            gameDuration: 604800,
-            fee: 10,
-            maxTicketsPerPlayer: 50,
-            isActive: true
-        });
+        gameConfigs[5] =
+            GameConfig({ticketPrice: 1 ether, gameDuration: 604800, fee: 10, maxTicketsPerPlayer: 50, isActive: true});
 
         // System parameters
         systemParams["minTicketPrice"] = 0.01 ether;
@@ -82,27 +67,15 @@ contract ConfigManager is Initializable, OwnableUpgradeable {
      */
     function updateGameConfig(
         uint8 gameType,
-        uint ticketPrice,
-        uint gameDuration,
+        uint256 ticketPrice,
+        uint256 gameDuration,
         uint8 fee,
-        uint maxTicketsPerPlayer
+        uint256 maxTicketsPerPlayer
     ) external onlyOwner {
-        require(
-            ticketPrice >= systemParams["minTicketPrice"],
-            "Ticket price too low"
-        );
-        require(
-            ticketPrice <= systemParams["maxTicketPrice"],
-            "Ticket price too high"
-        );
-        require(
-            gameDuration >= systemParams["minGameDuration"],
-            "Game duration too short"
-        );
-        require(
-            gameDuration <= systemParams["maxGameDuration"],
-            "Game duration too long"
-        );
+        require(ticketPrice >= systemParams["minTicketPrice"], "Ticket price too low");
+        require(ticketPrice <= systemParams["maxTicketPrice"], "Ticket price too high");
+        require(gameDuration >= systemParams["minGameDuration"], "Game duration too short");
+        require(gameDuration <= systemParams["maxGameDuration"], "Game duration too long");
         require(fee <= 20, "Fee too high");
         require(maxTicketsPerPlayer > 0, "Max tickets must be greater than 0");
 
@@ -114,13 +87,7 @@ contract ConfigManager is Initializable, OwnableUpgradeable {
             isActive: gameConfigs[gameType].isActive
         });
 
-        emit GameConfigUpdated(
-            gameType,
-            ticketPrice,
-            gameDuration,
-            fee,
-            maxTicketsPerPlayer
-        );
+        emit GameConfigUpdated(gameType, ticketPrice, gameDuration, fee, maxTicketsPerPlayer);
     }
 
     /**
@@ -133,11 +100,8 @@ contract ConfigManager is Initializable, OwnableUpgradeable {
     /**
      * @dev 시스템 파라미터 업데이트
      */
-    function updateSystemParam(
-        string memory param,
-        uint value
-    ) external onlyOwner {
-        uint oldValue = systemParams[param];
+    function updateSystemParam(string memory param, uint256 value) external onlyOwner {
+        uint256 oldValue = systemParams[param];
         systemParams[param] = value;
         emit SystemParamUpdated(param, oldValue, value);
     }
@@ -145,10 +109,7 @@ contract ConfigManager is Initializable, OwnableUpgradeable {
     /**
      * @dev 컨트랙트 주소 업데이트
      */
-    function updateContractAddress(
-        string memory contractName,
-        address newAddress
-    ) external onlyOwner {
+    function updateContractAddress(string memory contractName, address newAddress) external onlyOwner {
         require(newAddress != address(0), "Invalid contract address");
         address oldAddress = contractAddresses[contractName];
         contractAddresses[contractName] = newAddress;
@@ -158,25 +119,21 @@ contract ConfigManager is Initializable, OwnableUpgradeable {
     /**
      * @dev 게임 설정 조회
      */
-    function getGameConfig(
-        uint8 gameType
-    ) external view returns (GameConfig memory) {
+    function getGameConfig(uint8 gameType) external view returns (GameConfig memory) {
         return gameConfigs[gameType];
     }
 
     /**
      * @dev 시스템 파라미터 조회
      */
-    function getSystemParam(string memory param) external view returns (uint) {
+    function getSystemParam(string memory param) external view returns (uint256) {
         return systemParams[param];
     }
 
     /**
      * @dev 컨트랙트 주소 조회
      */
-    function getContractAddress(
-        string memory contractName
-    ) external view returns (address) {
+    function getContractAddress(string memory contractName) external view returns (address) {
         return contractAddresses[contractName];
     }
 }
