+++ contracts/shared/interfaces/ICryptolottoStatsAggregator.sol
@@ -7,22 +7,20 @@ pragma solidity ^0.8.19;
 interface ICryptolottoStatsAggregator {
     function newWinner(
         address winner,
-        uint game,
-        uint players,
-        uint amount,
+        uint256 game,
+        uint256 players,
+        uint256 amount,
         uint8 gameType,
-        uint winnerIndex
+        uint256 winnerIndex
     ) external;
 
-    function getWinnerCount(address player) external view returns (uint);
+    function getWinnerCount(address player) external view returns (uint256);
 
-    function getTotalWinnings(address player) external view returns (uint);
+    function getTotalWinnings(address player) external view returns (uint256);
 
-    function getGameWinner(uint game) external view returns (address);
+    function getGameWinner(uint256 game) external view returns (address);
 
-    function getPlayerStats(
-        address player
-    ) external view returns (uint wins, uint totalWon, bool hasWon);
+    function getPlayerStats(address player) external view returns (uint256 wins, uint256 totalWon, bool hasWon);
 
-    function getTopWinners(uint count) external view returns (address[] memory);
+    function getTopWinners(uint256 count) external view returns (address[] memory);
 }
